requirements.txt
azure-core==1.30.1
azure-functions
azure-storage-blob==12.19.1
backcall==0.2.0
click==8.1.7
cssselect==1.2.0
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.11
greenlet==3.0.3
install==1.3.5
ipython==8.12.3
isodate==0.6.1
jieba3k==0.35.1
joblib==1.4.0
lxml==5.2.1
lxml_html_clean==0.1.1
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
psycopg2==2.9.9
pyarrow==15.0.2
pytz==2024.1
regex==2023.12.25
requests-file==2.0.0
sgmllib3k==1.0.0
SQLAlchemy==2.0.20
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.2
tzdata==2024.1
yarg==0.1.9
